chore(riscv): split test and source code
