+++ .gitattributes
@@ -1,1 +1,3 @@
 *duktape_ee.bin filter=lfs diff=lfs merge=lfs -text
+*libpvm.a filter=lfs diff=lfs merge=lfs -text
+*pvm_structs.bin filter=lfs diff=lfs merge=lfs -text
